requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
Django==4.2.5
django-crispy-forms==2.0
django-localflavor==4.0
python-decouple==3.8
python-dotenv==1.0.0
python-stdnum==1.19
sqlparse==0.4.4
tzdata==2023.3
